cliplugins: pure move of enable-syncer code
