+++ pkg/cliplugins/workload/plugin/sync_test.go
